# protocol/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AudienceScore Pilot API",
    "version": "0.2.0-pilot",
    "description": "Pilot deployment, pre-cryptographic-audit. The independent cryptographic review and per-vertical legal review gates remain open; the pilot ledger may be reset and receipts re-issued after the audit. Reading scores requires no authentication. Pilot receipts, review events, and score manifests carry env=\"pilot\" in the signed body.",
    "license": {
      "name": "CC BY 4.0",
      "identifier": "CC-BY-4.0"
    }
  },
  "servers": [
    {
      "url": "https://mcp.audiencescore.org",
      "description": "Hosted pilot read API: /mcp and /v0/scores/*"
    },
    {
      "url": "https://{issuerWriteHost}",
      "description": "Operator-deployed issuer/write API running reference-impl/src/pilot/server.js",
      "variables": {
        "issuerWriteHost": {
          "default": "api.audiencescore.org",
          "description": "Configure DNS before using this host"
        }
      }
    }
  ],
  "paths": {
...
    "/v0/reviews": {
      "post": {
        "operationId": "submitReview",
        "summary": "Submit a pilot review by presenting a signed receipt",
        "description": "The holder or their agent presents the signed pilot receipt JSON and a review. The server verifies the receipt, confirms it matches the pilot ledger, then spends the standing through the v0.2 store admission path.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubmitReviewRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pilot review accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubmitReviewResponse" }
              }
            }
          },
          "400": { "description": "Invalid, missing, unknown, or wrong-issuer receipt" },
          "409": { "description": "Receipt/standing already reviewed" }
        }
      }
    },
...
    "/v1/transactions": {
      "post": {
        "operationId": "ingestTransaction",
        "summary": "Ingest a normalized partner transaction",
        "description": "Partner write endpoint. The request body must be signed by the partner Ed25519 key using x-as-timestamp, x-as-nonce, and x-as-signature. The signature covers METHOD, PATH, timestamp, nonce, and sha256(body).",
        "parameters": [
          { "name": "x-as-partner-id", "in": "header", "required": true, "schema": { "type": "string" } },
          { "name": "x-as-timestamp", "in": "header", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "x-as-nonce", "in": "header", "required": true, "schema": { "type": "string" } },
          { "name": "x-as-signature", "in": "header", "required": true, "schema": { "type": "string" }, "description": "Lowercase-hex Ed25519 signature." }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NormalizedTransaction" }
            }
          }
        },
        "responses": {
          "201": { "description": "Minted one receipt" },
          "200": { "description": "Corroborated, reversed, conflicted, or ambiguous" },
          "401": { "description": "Missing, bad, stale, or replayed partner signature" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Receipt": {
        "type": "object",
        "required": ["spec", "receipt_id", "issuer", "holder", "role", "offering", "level", "event", "issued_at", "prev", "env", "coattest", "sig"],
        "properties": {
          "spec": { "const": "as/0.2a" },
          "receipt_id": { "type": "string" },
          "issuer": { "type": "string" },
          "holder": { "type": "string", "description": "Pseudonymous holder binding; never plaintext customer identity" },
          "role": { "enum": ["participant", "payer"] },
          "offering": { "type": "string" },
          "level": { "type": "integer", "minimum": 1, "maximum": 4 },
          "event": { "type": "string" },
          "issued_at": { "type": "string", "format": "date-time" },
          "prev": { "type": ["string", "null"] },
          "env": { "const": "pilot" },
          "coattest": { "type": "array", "items": { "type": "string" } },
          "sig": { "type": "string" }
        }
      },
      "SubmitReviewRequest": {
        "type": "object",
        "required": ["receipt", "review"],
        "properties": {
          "receipt": { "$ref": "#/components/schemas/Receipt" },
          "review": {
            "type": "object",
            "required": ["overall"],
            "properties": {
              "overall": { "type": "integer", "minimum": 1, "maximum": 5 },
              "facets": { "type": "object", "additionalProperties": { "type": "integer", "minimum": 1, "maximum": 5 } },
              "text": { "type": ["string", "null"] }
            }
          }
        }
      },
      "SubmitReviewResponse": {
        "type": "object",
        "required": ["env", "reviewId", "reviewClass", "event"],
        "properties": {
          "env": { "const": "pilot" },
          "reviewId": { "type": "string" },
          "reviewClass": { "type": "string" },
          "event": { "type": "object" }
        }
      },
      "NormalizedTransaction": {
        "type": "object",
        "required": ["offering", "amountCents", "currency", "kind"],
        "properties": {
          "issuerId": { "type": "string" },
          "connectedAccountRef": { "type": "string" },
          "offering": { "type": "string", "examples": ["field-elevate-demo@v1"] },
          "rail": { "type": "string", "examples": ["stripe", "square", "manual"] },
          "processorTxnId": { "type": "string" },
          "amountCents": { "type": "integer", "minimum": 1 },
          "currency": { "type": "string", "examples": ["usd"] },
          "occurredAt": { "type": "string", "format": "date-time" },
          "customerEmail": { "type": "string", "format": "email" },
          "role": { "enum": ["participant", "payer"] },
          "kind": { "enum": ["transaction", "reversal"] }
        }
      },
...
    }
  }
}